refactor: 把表情包目录配置从编辑器剥离为独立模块
四个转写 CLI 只需要 .env 里的 STICKER_DIR 一项配置，却写成
`from edit import get_default_sticker_dir`，于是每个转写入口都会连带导入
整个 MAWE 生成器和它依赖的 Rust .ReaPeaks 波形内核。

先把 load_env / get_default_sticker_dir 移到叶子模块 maw.stickers（只依赖
标准库和 maw.app_paths），CLI 改用它，edit.py 保留同名再导出以兼容
server-editor 的调用；MAW.spec 同步把新模块复制进 local-runtime 镜像，
并加契约测试守住本地转写不再导入编辑器栈。
