+++ MAW.spec
@@ -74,6 +74,7 @@ datas = [
     (str(ROOT / "maw" / "qwen_audio.py"), "local-runtime/maw"),
     (str(ROOT / "maw" / "colors.py"), "local-runtime/maw"),
     (str(ROOT / "maw" / "speaker.py"), "local-runtime/maw"),
+    (str(ROOT / "maw" / "stickers.py"), "local-runtime/maw"),
     (str(ROOT / "maw" / "__init__.py"), "ocr-runtime/maw"),
     (str(ROOT / "maw" / "console.py"), "ocr-runtime/maw"),
     (str(ROOT / "maw" / "media.py"), "ocr-runtime/maw"),
